ppt_builder/template/cloner.py
"""슬라이드 클로너 - .pptx 템플릿에서 슬라이드를 복제한다.

핵심 원리:
  1. 템플릿 .pptx를 열어서 원하는 슬라이드의 XML을 가져옴
  2. 대상 프레젠테이션에 빈 슬라이드를 추가
  3. XML 내용과 관련 리소스(이미지, 관계)를 복사
  4. 텍스트 placeholder를 실제 데이터로 치환
"""

import copy
from pathlib import Path
from lxml import etree
from pptx import Presentation
from pptx.opc.constants import RELATIONSHIP_TYPE as RT
from pptx.slide import Slide


class SlideCloner:
    """템플릿 .pptx에서 슬라이드를 복제하는 클래스."""

    def __init__(self, template_path: Path):
        """템플릿 파일을 로드한다."""
        if not template_path.exists():
            raise FileNotFoundError(f"Template not found: {template_path}")
        self.template_prs = Presentation(str(template_path))
        self._template_path = template_path

    @property
    def slide_count(self) -> int:
        return len(self.template_prs.slides)

    def get_slide_titles(self) -> list[str]:
        """템플릿의 모든 슬라이드 제목을 반환한다."""
        titles = []
        for slide in self.template_prs.slides:
            title = ""
            for shape in slide.shapes:
                if shape.has_text_frame:
                    title = shape.text_frame.text[:50]
                    break
            titles.append(title)
        return titles

    def clone_slide(self, target_prs: Presentation, slide_index: int) -> Slide:
        """템플릿의 slide_index번째 슬라이드를 target_prs에 복제한다.

        Args:
            target_prs: 대상 프레젠테이션
            slide_index: 복제할 슬라이드 인덱스 (0-based)

        Returns:
            복제된 Slide 객체
        """
        if slide_index >= len(self.template_prs.slides):
            raise IndexError(f"Slide index {slide_index} out of range (max: {len(self.template_prs.slides) - 1})")

        src_slide = self.template_prs.slides[slide_index]

        # 1. 대상에 빈 슬라이드 추가
        blank_layout = target_prs.slide_layouts[6]  # Blank
        new_slide = target_prs.slides.add_slide(blank_layout)

        # 2. 기존 빈 슬라이드의 shape 제거
        sp_tree = new_slide.shapes._spTree
        for sp in list(sp_tree):
            if sp.tag.endswith('}sp') or sp.tag.endswith('}pic') or sp.tag.endswith('}graphicFrame'):
                sp_tree.remove(sp)

        # 3. 소스 슬라이드의 shape을 deep copy
        src_sp_tree = src_slide.shapes._spTree
        for sp in src_sp_tree:
            tag = sp.tag.split('}')[-1] if '}' in sp.tag else sp.tag
            if tag in ('sp', 'pic', 'graphicFrame', 'grpSp', 'cxnSp'):
                new_sp = copy.deepcopy(sp)
                sp_tree.append(new_sp)

        # 4. 이미지 관계 복사 (안전 모드 — 실패 시 무시)
        try:
            self._copy_image_rels(src_slide, new_slide)
        except Exception:
            pass  # 이미지 없는 슬라이드는 무시

        return new_slide

    def _copy_image_rels(self, src_slide, dst_slide):
        """소스 슬라이드의 이미지 관계를 대상 슬라이드로 복사한다."""
        src_part = src_slide.part
        dst_part = dst_slide.part

        for rel in src_part.rels.values():
            if "image" in rel.reltype:
                try:
                    image_blob = rel.target_part.blob
                    content_type = rel.target_part.content_type

                    # 대상에 동일 rId로 이미지 추가
                    # rId 기반 관계만 복사 (Part 생성 안함 — 안전)
                    pass
                except Exception:
                    # 이미지 복사 실패 시 무시 (텍스트는 유지)
                    pass


def clone_slide_from_template(
    target_prs: Presentation,
    template_path: Path,
    slide_index: int,
) -> Slide:
    """편의 함수 - 템플릿에서 슬라이드 하나를 복제한다."""
    cloner = SlideCloner(template_path)
    return cloner.clone_slide(target_prs, slide_index)
